Separated context graph type into multiple files.
