+++ deep_causality/src/types/context_types/context_graph/contextuable_graph.rs
@@ -1,66 +1,7 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) "2023" . Marvin Hansen <marvin.hansen@gmail.com> All rights reserved.
 
-use std::fmt::{Debug, Display, Formatter};
-
-use ultragraph::prelude::*;
-
-use crate::errors::ContextIndexError;
-use crate::prelude::{Contextoid, ContextuableGraph, Datable, Identifiable, RelationKind, SpaceTemporal, Spatial, Temporal};
-
-pub struct Context<'l, D, S, T, ST>
-    where
-        D: Datable,
-        S: Spatial,
-        T: Temporal,
-        ST: SpaceTemporal,
-{
-    id: u64,
-    name: &'l str,
-    graph: UltraGraph<Contextoid<D, S, T, ST>>,
-}
-
-
-impl<'l, D, S, T, ST> Context<'l, D, S, T, ST>
-    where
-        D: Datable,
-        S: Spatial,
-        T: Temporal,
-        ST: SpaceTemporal,
-{
-    /// Creates a new context with the given node capacity.
-    pub fn with_capacity(
-        id: u64,
-        name: &'l str,
-        capacity: usize,
-    )
-        -> Self
-    {
-        Self {
-            id,
-            name,
-            graph: ultragraph::new_with_matrix_storage(capacity),
-        }
-    }
-
-    /// Returns the name of the context.
-    pub fn name(&self) -> &str {
-        self.name
-    }
-}
-
-impl<'l, D, S, T, ST> Identifiable for Context<'l, D, S, T, ST>
-    where
-        D: Datable,
-        S: Spatial,
-        T: Temporal,
-        ST: SpaceTemporal,
-{
-    /// Returns the id of the context.
-    fn id(&self) -> u64 {
-        self.id
-    }
-}
+use super::*;
 
 impl<'l, D, S, T, ST> ContextuableGraph<'l, D, S, T, ST> for Context<'l, D, S, T, ST>
     where
@@ -219,47 +160,5 @@ impl<'l, D, S, T, ST> ContextuableGraph<'l, D, S, T, ST> for Context<'l, D, S, T
         -> usize
     {
         self.graph.number_edges()
-    }
-}
-
-impl<'l, D, S, T, ST> Context<'l, D, S, T, ST>
-    where
-        D: Datable,
-        S: Spatial,
-        T: Temporal,
-        ST: SpaceTemporal,
-{
-    fn format(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f,
-               "Context: id: {}, name: {}, node_count: {}, edge_count: {}",
-               self.id,
-               self.name,
-               self.node_count(),
-               self.edge_count(),
-        )
-    }
-}
-
-impl<'l, D, S, T, ST> Debug for Context<'l, D, S, T, ST>
-    where
-        D: Datable,
-        S: Spatial,
-        T: Temporal,
-        ST: SpaceTemporal,
-{
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        self.format(f)
-    }
-}
-
-impl<'l, D, S, T, ST> Display for Context<'l, D, S, T, ST>
-    where
-        D: Datable,
-        S: Spatial,
-        T: Temporal,
-        ST: SpaceTemporal,
-{
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        self.format(f)
     }
 }
