+++ deep_causality/src/types/context_types/context_graph/debug.rs
@@ -1,0 +1,47 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) "2023" . Marvin Hansen <marvin.hansen@gmail.com> All rights reserved.
+use std::fmt::{Debug, Display, Formatter};
+
+use crate::prelude::{Context, ContextuableGraph, Datable, SpaceTemporal, Spatial, Temporal};
+
+impl<'l, D, S, T, ST> Context<'l, D, S, T, ST>
+    where
+        D: Datable,
+        S: Spatial,
+        T: Temporal,
+        ST: SpaceTemporal,
+{
+    fn format(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f,
+               "Context: id: {}, name: {}, node_count: {}, edge_count: {}",
+               self.id,
+               self.name,
+               self.node_count(),
+               self.edge_count(),
+        )
+    }
+}
+
+impl<'l, D, S, T, ST> Debug for Context<'l, D, S, T, ST>
+    where
+        D: Datable,
+        S: Spatial,
+        T: Temporal,
+        ST: SpaceTemporal,
+{
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        self.format(f)
+    }
+}
+
+impl<'l, D, S, T, ST> Display for Context<'l, D, S, T, ST>
+    where
+        D: Datable,
+        S: Spatial,
+        T: Temporal,
+        ST: SpaceTemporal,
+{
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        self.format(f)
+    }
+}
