+++ deep_causality/src/types/context_types/mod.rs
@@ -2,7 +2,7 @@
 // Copyright (c) "2023" . Marvin Hansen <marvin.hansen@gmail.com> All rights reserved.
 
 pub mod node_types;
-pub mod context;
 pub mod time_scale;
 pub mod contextoid;
 pub mod relation_kind;
+pub mod context_graph;
